requirements.txt
# RL ve Gym
gymnasium==0.29.1
stable-baselines3==2.2.1
torch==2.1.0
tensorboard==2.15.1

# Data handling
pandas==2.1.3
numpy==1.24.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1 